+++ app/reducers/index.js
@@ -2,6 +2,7 @@ import { combineReducers } from 'redux';
 import user from 'reducers/user';
 import topic from 'reducers/topic';
 import message from 'reducers/message';
+import sheet from 'reducers/sheet';
 import { routerReducer as routing } from 'react-router-redux';
 
 // Combine reducers with routeReducer which keeps track of
@@ -10,6 +11,7 @@ const rootReducer = combineReducers({
   user,
   topic,
   message,
+  sheet,
   routing
 });
